[tempo-distributed] Ensure that Admin Client is configured with relevant details.
Previously, the Admin Client section of the chart for Grafana
Enterprise Traces was only configured should MinIO be enabled.

These changes:
* Add a new `storage.admin` section for configuring the client
  backend storage type and bucket details.
* Use the configured MinIO bucket details without hardcoding the
  values (they get picked up from the `minio` section).
* Ensures the main storage configuration section includes a
  configuration section should MinIO be enabled.
* Finally, the README has been updated to document the new
  admin storage configuration section.

Additionally, the MinIO chart has been updated to the latest
vanilla chart. This ensures that MinIO actually starts and is
configurable (the previous version did not).

This commit fixes #2111, #2112.

Signed-off-by: Heds Simons <hedley.simons@grafana.com>